Set up TypeScript
B changing the `index.js` to `index.tsx` and running `next` we can have
it generate us a `tsconfig.json` (and spit out the `npm install`s we
need for types). The change I have made to the default config is to turn
on `strict` mode because I want all the help I can get from the
compiler.

n.b. This is a new project, so starting strict is reasonable. On an
existing project migrating to TypeScript it would likely be necessary to
gradually migrate to this level of strictness depending on how
"JavaScript-y" it was.