Move home page content to `containers` directory
Due to quirks of how Next discovers pages, we don't want tests in the
`pages` directory. However, it is my strong preference to colocate tests
with the files they test. Therefore, I'm using the `containers` folder
to actually put pages in. The files in `pages` will be a mirrored
structure of `containers`.