Set up Jest for tests
In addition to Jest, adds [Testing Library](https://testing-library.com)
which I consider a must have tool for testing.

Adding a babel config is necessary to ensure that Jest and specifically
`babel-test` can handle TypeScript files. Some of the junk in the
`jest.config.js` is premature or potentially unnecessary. I copy pasted
it from another project, as is tradition.