Limit decimal input to only valid values
While this may not be a completely robust implementation it mostly does
the job. Allowes typing in valid sequences of charachters which would
form a decimal number. For example, the follwoing would all be allowed:
0
1
10
100
10.
10.1
10.12

The following would not be allowed:
asdf
10.x
10..1

Because this is backed by an HTML input, the value is a string.
Validating it as a number needs to happen elsewhere.